Move MgVideo to _video.py to fix recursive import
